chore: rename nearest_trajectories_pose_re.. to nearest_trajectories
