+++ src/nearest_trajectories.rs
